fix: 타입 변경 및 현재 불필요한 코드 제거

+++ src/main/java/com/stocking/modules/account/AccountController.java
@@ -17,7 +17,7 @@ public class AccountController {
     private final AccountService accountService;
 
     @GetMapping("/login")
-    public ResponseEntity<Object> login(Integer uuid, Error error) {
+    public ResponseEntity<Object> login(Long uuid, Error error) {
         Account account = accountService.findByUuid(uuid);
         if (account == null) {
             return null;
